+++ src-tauri/src/lib.rs
@@ -39,6 +39,7 @@ pub fn run() {
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_fs::init())
         .plugin(tauri_plugin_opener::init())
+        .plugin(tauri_plugin_http::init())
         .manage(EngineController::default())
         .manage(Arc::new(ConnectionManager::default()))
         .manage(MeshSession::default())